chore: move credentialsImage in the appropriate path
